refactor: import baseUrl 설정 및 product 도메인 관련 api 파일 위치 이동
