+++ src/components/features/product/api/deleteCartItem.ts
@@ -1,4 +1,4 @@
-import { baseAPI } from './baseAPI';
+import { baseAPI } from '@/api/baseAPI';
 
 export async function deleteCartItem(cartId: string) {
   await baseAPI({
