+++ src/context/CartProvider.tsx
@@ -1,3 +1,5 @@
+import { getShoppingCartList } from '@/components/features/product/api/getShoppingCartList';
+import { useShopErrorContext } from '@/shop/context/useShopErrorContext';
 import {
   createContext,
   useCallback,
@@ -6,8 +8,6 @@ import {
   useState,
 } from 'react';
 import { Cart } from '../components/features/product/type';
-import { getShoppingCartData } from '../api/getShoppingCartData';
-import { useShopErrorContext } from '../shop/context/useShopErrorContext';
 
 export const CartContext = createContext<{
   cartList: Cart[];
@@ -21,7 +21,7 @@ function CartProvider({ children }: { children: React.ReactNode }) {
 
   const getShoppingCartDataHandler = useCallback(async () => {
     try {
-      const cartsData = await getShoppingCartData();
+      const cartsData = await getShoppingCartList();
       if (cartsData) setCartList(cartsData);
       handleErrorFalse();
     } catch (e) {
